# packages/contracts/openapi/kitchenflow-v1.json
{
  "components": {
    "schemas": {
...
      "CreateLotRequest": {
        "description": "Creates one manually entered, user-owned inventory lot.",
        "properties": {
          "customLocation": {
            "type": [
              "null",
              "string"
            ]
          },
          "notes": {
            "type": [
              "null",
              "string"
            ]
          },
          "packageState": {
            "anyOf": [
              {
                "enum": [
                  "Sealed",
                  "Opened",
                  "Unknown"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "printedExpirationDate": {
            "format": "date",
            "type": [
              "null",
              "string"
            ]
          },
          "productName": {
            "type": "string"
          },
          "quantity": {
            "$ref": "#/components/schemas/QuantityRequest"
          },
          "storageLocation": {
            "enum": [
              "Pantry",
              "Refrigerator",
              "Freezer",
              "Other"
            ],
            "type": "string"
          }
        },
        "required": [
          "productName",
          "quantity",
          "storageLocation",
          "customLocation",
          "packageState",
          "printedExpirationDate",
          "notes"
        ],
        "type": "object"
      },
...
      "ListLotsResponse": {
        "description": "Returns a cursor-paginated owner-scoped inventory list.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/LotResponse"
            },
            "type": "array"
          },
          "nextCursor": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "required": [
          "items",
          "nextCursor"
        ],
        "type": "object"
      },
...
      "LotResponse": {
        "description": "Returns the current representation of an inventory lot.",
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "customLocation": {
            "type": [
              "null",
              "string"
            ]
          },
          "lotId": {
            "format": "uuid",
            "type": "string"
          },
          "notes": {
            "type": [
              "null",
              "string"
            ]
          },
          "packageState": {
            "type": [
              "null",
              "string"
            ]
          },
          "printedExpirationDate": {
            "format": "date",
            "type": [
              "null",
              "string"
            ]
          },
          "productId": {
            "format": "uuid",
            "type": "string"
          },
          "productName": {
            "type": "string"
          },
          "quantity": {
            "$ref": "#/components/schemas/QuantityResponse"
          },
          "storageLocation": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "version": {
            "type": "string"
          }
        },
        "required": [
          "lotId",
          "productId",
          "productName",
          "quantity",
          "storageLocation",
          "customLocation",
          "packageState",
          "printedExpirationDate",
          "notes",
          "version",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      },
...
      "ProblemDetails": {
        "properties": {
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "errorCode": {
            "description": "Stable machine-readable KitchenFlow error code.",
            "type": "string"
          },
          "errors": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "description": "Field-level validation messages keyed by request field name.",
            "type": "object"
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": "integer"
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "traceId": {
            "description": "Correlation identifier suitable for support without exposing private request data.",
            "type": "string"
          },
          "type": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "required": [
          "errorCode",
          "traceId"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "kitchenflowSession": {
        "description": "Backend-managed, secure HttpOnly session cookie. Browser code must not read or store OIDC tokens.",
        "in": "cookie",
        "name": "__Host-kitchenflow-session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "license": {
      "name": "PolyForm Noncommercial License 1.0.0",
      "url": "https://polyformproject.org/licenses/noncommercial/1.0.0/"
    },
    "title": "KitchenFlow.Api | v1",
    "version": "1.0.0"
  },
  "openapi": "3.1.1",
  "paths": {
...
    "/api/v1/inventory/lots": {
      "get": {
        "operationId": "getApiV1InventoryLots",
        "parameters": [
          {
            "in": "query",
            "name": "pageSize",
            "schema": {
              "format": "int32",
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ]
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "storageLocation",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListLotsResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "invalidCursor": {
                    "summary": "The cursor is invalid or has been tampered with.",
                    "value": {
                      "detail": "The cursor is invalid.",
                      "errorCode": "invalid_cursor",
                      "status": 400,
                      "title": "Bad Request",
                      "traceId": "00-00000000000000000000000000000000-0000000000000000-00",
                      "type": "about:blank"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "authenticationFailure": {
                    "summary": "The backend-managed session is absent or invalid.",
                    "value": {
                      "detail": "Authentication is required.",
                      "errorCode": "authentication_required",
                      "status": 401,
                      "title": "Unauthorized",
                      "traceId": "00-00000000000000000000000000000000-0000000000000000-00",
                      "type": "about:blank"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Authentication is required."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "unexpectedFailure": {
                    "summary": "An unexpected error was safely redacted.",
                    "value": {
                      "detail": "An unexpected error occurred.",
                      "errorCode": "unexpected_error",
                      "status": 500,
                      "title": "Internal Server Error",
                      "traceId": "00-00000000000000000000000000000000-0000000000000000-00",
                      "type": "about:blank"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "An unexpected server error occurred."
          }
        },
        "security": [
          {
            "kitchenflowSession": []
          }
        ],
        "summary": "GET /api/v1/inventory/lots",
        "tags": [
          "KitchenFlow.Api"
        ]
      },
      "post": {
        "operationId": "postApiV1InventoryLots",
        "parameters": [
          {
            "description": "Required CSRF token issued by GET /api/v1/session.",
            "in": "header",
            "name": "X-CSRF-TOKEN",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Required UUID key for semantic create or adjustment replay.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "availabilityLot": {
                  "summary": "A manually entered qualitative availability lot.",
                  "value": {
                    "customLocation": null,
                    "notes": null,
                    "packageState": null,
                    "printedExpirationDate": null,
                    "productName": "Fresh herbs",
                    "quantity": {
                      "availabilityState": "Low",
                      "measuredValue": null,
                      "unit": null
                    },
                    "storageLocation": "Refrigerator"
                  }
                },
                "measuredLot": {
                  "summary": "A manually entered measured pantry lot.",
                  "value": {
                    "customLocation": null,
                    "notes": null,
                    "packageState": "Sealed",
                    "printedExpirationDate": "2026-12-31",
                    "productName": "Red lentils",
                    "quantity": {
                      "availabilityState": null,
                      "measuredValue": 500,
                      "unit": "Gram"
                    },
                    "storageLocation": "Pantry"
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/CreateLotRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "examples": {
                  "completedCreate": {
                    "summary": "A newly completed create command.",
                    "value": {
                      "createdAt": "2026-07-31T00:00:00Z",
                      "customLocation": null,
                      "lotId": "00000000-0000-0000-0000-000000000001",
                      "notes": null,
                      "packageState": "Sealed",
                      "printedExpirationDate": "2026-12-31",
                      "productId": "00000000-0000-0000-0000-000000000002",
                      "productName": "Red lentils",
                      "quantity": {
                        "availabilityState": null,
                        "measuredValue": 500,
                        "unit": "Gram"
                      },
                      "storageLocation": "Pantry",
                      "updatedAt": "2026-07-31T00:00:00Z",
                      "version": "opaque-create-version"
                    }
                  },
                  "completedCreateReplay": {
                    "summary": "The exact semantic response returned for a completed create replay.",
                    "value": {
                      "createdAt": "2026-07-31T00:00:00Z",
                      "customLocation": null,
                      "lotId": "00000000-0000-0000-0000-000000000001",
                      "notes": null,
                      "packageState": "Sealed",
                      "printedExpirationDate": "2026-12-31",
                      "productId": "00000000-0000-0000-0000-000000000002",
                      "productName": "Red lentils",
                      "quantity": {
                        "availabilityState": null,
                        "measuredValue": 500,
                        "unit": "Gram"
                      },
                      "storageLocation": "Pantry",
                      "updatedAt": "2026-07-31T00:00:00Z",
                      "version": "opaque-create-version"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/LotResponse"
                }
              }
            },
            "description": "Created",
            "headers": {
              "ETag": {
                "description": "Opaque current version required by If-Match on subsequent mutations.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "fieldValidationFailure": {
                    "summary": "A required header or request field is malformed.",
                    "value": {
                      "detail": "A UUID Idempotency-Key header is required.",
                      "errorCode": "validation_failed",
                      "errors": {
                        "Idempotency-Key": [
                          "A UUID Idempotency-Key header is required."
                        ]
                      },
                      "status": 400,
                      "title": "Bad Request",
                      "traceId": "00-00000000000000000000000000000000-0000000000000000-00",
                      "type": "about:blank"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "authenticationFailure": {
                    "summary": "The backend-managed session is absent or invalid.",
                    "value": {
                      "detail": "Authentication is required.",
                      "errorCode": "authentication_required",
                      "status": 401,
                      "title": "Unauthorized",
                      "traceId": "00-00000000000000000000000000000000-0000000000000000-00",
                      "type": "about:blank"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Authentication is required."
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict"
          },
          "415": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "unsupportedMediaType": {
                    "summary": "The request content type is unsupported.",
                    "value": {
                      "detail": "The request content type is not supported.",
                      "errorCode": "unsupported_media_type",
                      "status": 415,
                      "title": "Unsupported Media Type",
                      "traceId": "00-00000000000000000000000000000000-0000000000000000-00",
                      "type": "about:blank"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The request content type is not supported."
          },
          "422": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "invalidQuantity": {
                    "summary": "A measured and availability quantity cannot be supplied together.",
                    "value": {
                      "detail": "Quantity must use exactly one mode.",
                      "errorCode": "domain_rule_violated",
                      "errors": {
                        "quantity": [
                          "Quantity must use exactly one mode."
                        ]
                      },
                      "status": 422,
                      "title": "Unprocessable Entity",
                      "traceId": "00-00000000000000000000000000000000-0000000000000000-00",
                      "type": "about:blank"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "rateLimitFailure": {
                    "summary": "The bounded mutation rate was exceeded.",
                    "value": {
                      "detail": "The request rate limit was exceeded.",
                      "errorCode": "rate_limit_exceeded",
                      "status": 429,
                      "title": "Too Many Requests",
                      "traceId": "00-00000000000000000000000000000000-0000000000000000-00",
                      "type": "about:blank"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The mutation rate limit was exceeded."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "unexpectedFailure": {
                    "summary": "An unexpected error was safely redacted.",
                    "value": {
                      "detail": "An unexpected error occurred.",
                      "errorCode": "unexpected_error",
                      "status": 500,
                      "title": "Internal Server Error",
                      "traceId": "00-00000000000000000000000000000000-0000000000000000-00",
                      "type": "about:blank"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "An unexpected server error occurred."
          }
        },
        "security": [
          {
            "kitchenflowSession": []
          }
        ],
        "summary": "POST /api/v1/inventory/lots",
        "tags": [
          "KitchenFlow.Api"
        ]
      }
    },
...
    "/api/v1/inventory/lots/{lotId}/history": {
      "get": {
        "operationId": "getApiV1InventoryLotsLotIdHistory",
        "parameters": [
          {
            "in": "path",
            "name": "lotId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LotHistoryResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "authenticationFailure": {
                    "summary": "The backend-managed session is absent or invalid.",
                    "value": {
                      "detail": "Authentication is required.",
                      "errorCode": "authentication_required",
                      "status": 401,
                      "title": "Unauthorized",
                      "traceId": "00-00000000000000000000000000000000-0000000000000000-00",
                      "type": "about:blank"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Authentication is required."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "unexpectedFailure": {
                    "summary": "An unexpected error was safely redacted.",
                    "value": {
                      "detail": "An unexpected error occurred.",
                      "errorCode": "unexpected_error",
                      "status": 500,
                      "title": "Internal Server Error",
                      "traceId": "00-00000000000000000000000000000000-0000000000000000-00",
                      "type": "about:blank"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "An unexpected server error occurred."
          }
        },
        "security": [
          {
            "kitchenflowSession": []
          }
        ],
        "summary": "GET /api/v1/inventory/lots/{lotId}/history",
        "tags": [
          "KitchenFlow.Api"
        ]
      }
    },
    "/api/v1/inventory/lots/{lotId}/provenance": {
      "get": {
        "operationId": "getApiV1InventoryLotsLotIdProvenance",
        "parameters": [
          {
            "in": "path",
            "name": "lotId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LotProvenanceResponse"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "authenticationFailure": {
                    "summary": "The backend-managed session is absent or invalid.",
                    "value": {
                      "detail": "Authentication is required.",
                      "errorCode": "authentication_required",
                      "status": 401,
                      "title": "Unauthorized",
                      "traceId": "00-00000000000000000000000000000000-0000000000000000-00",
                      "type": "about:blank"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Authentication is required."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "unexpectedFailure": {
                    "summary": "An unexpected error was safely redacted.",
                    "value": {
                      "detail": "An unexpected error occurred.",
                      "errorCode": "unexpected_error",
                      "status": 500,
                      "title": "Internal Server Error",
                      "traceId": "00-00000000000000000000000000000000-0000000000000000-00",
                      "type": "about:blank"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "An unexpected server error occurred."
          }
        },
        "security": [
          {
            "kitchenflowSession": []
          }
        ],
        "summary": "GET /api/v1/inventory/lots/{lotId}/provenance",
        "tags": [
          "KitchenFlow.Api"
        ]
      }
    },
...
    "/api/v1/inventory/preparations/{batchId}": {
      "get": {
        "operationId": "getApiV1InventoryPreparationsBatchId",
        "parameters": [
          {
            "in": "path",
            "name": "batchId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreparationResponse"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "authenticationFailure": {
                    "summary": "The backend-managed session is absent or invalid.",
                    "value": {
                      "detail": "Authentication is required.",
                      "errorCode": "authentication_required",
                      "status": 401,
                      "title": "Unauthorized",
                      "traceId": "00-00000000000000000000000000000000-0000000000000000-00",
                      "type": "about:blank"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Authentication is required."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "unexpectedFailure": {
                    "summary": "An unexpected error was safely redacted.",
                    "value": {
                      "detail": "An unexpected error occurred.",
                      "errorCode": "unexpected_error",
                      "status": 500,
                      "title": "Internal Server Error",
                      "traceId": "00-00000000000000000000000000000000-0000000000000000-00",
                      "type": "about:blank"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "An unexpected server error occurred."
          }
        },
        "security": [
          {
            "kitchenflowSession": []
          }
        ],
        "summary": "GET /api/v1/inventory/preparations/{batchId}",
        "tags": [
          "KitchenFlow.Api"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Current KitchenFlow backend origin. Local development defaults to https://localhost:7443.",
      "url": "/"
    }
  ],
...
}